feat(web): remove the Inbox; surface org invitations in the org switcher
The global "Inbox" popover is gone. Pending cross-org invitations now render as
rows directly in the org switcher (left topbar) — each shows the org and an
Accept button (accepting joins + navigates there), with a decline affordance.
Its other feeds already have homes: join requests in Settings, release updates
in the floating release card.

Rebuilt the sidebar footer (account, invite, repo switcher, add connection,
settings, desktop link) without the Inbox row — the old footer component
bundled all of it together. Deleted the now-dead inbox feed, invitation-item,
inbox-release-item, and join-request-item.
