+++ apps/mesh/src/web/components/sidebar/footer/sidebar-footer.tsx
@@ -1,0 +1,129 @@
+import { AccountPopover } from "@/web/components/account-popover";
+import {
+  SidebarFooter,
+  SidebarMenu,
+  SidebarMenuButton,
+  SidebarMenuItem,
+  useSidebar,
+} from "@deco/ui/components/sidebar.tsx";
+import { Settings02, UserPlus01, ZapSquare } from "@untitledui/icons";
+import { useState } from "react";
+import { RepoSwitcher } from "@/web/components/sidebar/footer/repo-switcher";
+import { InviteMemberDialog } from "@/web/components/invite-member-dialog";
+import { AddConnectionDialog } from "@/web/views/virtual-mcp/add-connection-dialog";
+import { useProjectContext } from "@decocms/mesh-sdk";
+import { useNavigate } from "@tanstack/react-router";
+import { LinkedDesktopIndicator } from "@/web/components/header/linked-desktop-indicator";
+import { SidebarTopActions } from "@/web/components/sidebar/top-actions";
+
+function SettingsFullButton() {
+  const navigate = useNavigate();
+  const { org } = useProjectContext();
+  return (
+    <SidebarMenuButton
+      tooltip="Settings"
+      onClick={() =>
+        navigate({
+          to: "/$org/settings",
+          params: { org: org.slug },
+        })
+      }
+    >
+      <Settings02 />
+      <span>Settings</span>
+    </SidebarMenuButton>
+  );
+}
+
+/**
+ * Sidebar footer actions — org-wide entry points that aren't tied to a single
+ * agent: invite teammates, add a repo as an org-shared GitHub connection
+ * (available to every agent), and add any connection. Rendered as full-width
+ * rows above the account row.
+ */
+function SidebarExtraActions() {
+  const [connectionsOpen, setConnectionsOpen] = useState(false);
+  return (
+    <>
+      <SidebarMenu className="gap-0.5">
+        <SidebarMenuItem>
+          <InviteMemberDialog
+            trigger={
+              <SidebarMenuButton tooltip="Invite members">
+                <UserPlus01 />
+                <span>Invite members</span>
+              </SidebarMenuButton>
+            }
+          />
+        </SidebarMenuItem>
+        <SidebarMenuItem>
+          <RepoSwitcher />
+        </SidebarMenuItem>
+        <SidebarMenuItem>
+          <SidebarMenuButton
+            tooltip="Add connection"
+            onClick={() => setConnectionsOpen(true)}
+          >
+            <ZapSquare />
+            <span>Add connection</span>
+          </SidebarMenuButton>
+        </SidebarMenuItem>
+      </SidebarMenu>
+      <AddConnectionDialog
+        open={connectionsOpen}
+        onOpenChange={setConnectionsOpen}
+        mode="browse"
+      />
+    </>
+  );
+}
+
+/**
+ * Sidebar footer: org-wide actions + the account row. (The old global "Inbox"
+ * was removed — invitations now live in the org switcher, join requests in
+ * Settings, and release updates in the floating release card.)
+ */
+export function SidebarAccountFooter() {
+  const { state } = useSidebar();
+  const isCollapsed = state === "collapsed";
+
+  if (isCollapsed) {
+    return (
+      <SidebarFooter className="px-2 pb-3 gap-1">
+        <SidebarExtraActions />
+        <SidebarTopActions />
+        <SidebarMenu>
+          <SidebarMenuItem>
+            <div className="flex justify-center">
+              <LinkedDesktopIndicator />
+            </div>
+          </SidebarMenuItem>
+          <SidebarMenuItem>
+            <SettingsFullButton />
+          </SidebarMenuItem>
+          <SidebarMenuItem>
+            <AccountPopover />
+          </SidebarMenuItem>
+        </SidebarMenu>
+      </SidebarFooter>
+    );
+  }
+
+  return (
+    <SidebarFooter className="px-2 pb-3 gap-0.5">
+      <SidebarExtraActions />
+      <SidebarTopActions />
+      <SidebarMenu className="gap-0.5">
+        <SidebarMenuItem>
+          <LinkedDesktopIndicator variant="full" />
+        </SidebarMenuItem>
+        <SidebarMenuItem>
+          <SettingsFullButton />
+        </SidebarMenuItem>
+        <SidebarMenuItem>
+          <AccountPopover />
+        </SidebarMenuItem>
+      </SidebarMenu>
+    </SidebarFooter>
+  );
+}
